pyproject.toml
[tool.poetry]
name = "flask-gb"
version = "0.1.0"
description = ""
authors = ["Alexey Ryzhkov <aler32@mail.ru>"]
readme = "README.md"
packages = [{include = "flask_gb"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-sqlalchemy = "^3.0.3"
flask-login = "^0.6.2"
flask-migrate = "^4.0.4"
psycopg2-binary = "^2.9.5"
bcrypt-flask = "^1.0.2"
wtforms = "^3.0.1"
email-validator = "^1.3.1"
flask-wtf = "1.1.1"
flask-admin = "^1.6.1"
flask-combo-jsonapi = { git = "https://github.com/AdCombo/flask-combo-jsonapi.git" }
#combojsonapi = { git = "https://github.com/AdCombo/combojsonapi.git" }
apispec = "^6.3.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
